Teach tests knowledge of RMQ env var
Consume the environment variable if it exists, but fallback to a likely-good
default value.  Either way, perform an initial connection ("sanity") test to
verify the URI's validity -- if it fails, provide a hopefully
helpful/easy-to-spot error message for the dev.

For explicitness, the variable is named `RABBITMQ_INTEGRATION_TEST_URI`.  A
friendly shell reminder that to use it, can either set it per invocation:

    $ RABBITMQ_INTEGRATION_TEST_URI=amqp://u:p@rabbit... tox -e py

or export for the entire shell session:

    $ export RABBITMQ_INTEGRATION_TEST_URI=amqp://u:p@rabbit...
    $ tox -e py

Meanwhile, DRY out the fixtures a bit by utilizing the shared hierarchy.
