+++ funcx_endpoint/tox.ini
@@ -3,8 +3,10 @@ envlist = py{37,310}
 skip_missing_interpreters = true
 
 [testenv]
-usedevelop = true
+passenv =
+    RABBITMQ_INTEGRATION_TEST_URI
 extras = test
+usedevelop = true
 commands =
     coverage erase
     coverage run -m pytest {posargs}
